Refactor: Update project structure and README
This commit restructures the project to use a unified architecture and updates the README to reflect the new features and installation process.
