+++ logs/usa_app.log
@@ -1,0 +1,2 @@
+2025-10-24 19:23:04,414 INFO: Background monitor started
+2025-10-24 19:23:42,027 INFO: Background monitor started
