+++ static/js/app.js
@@ -1,103 +1,67 @@
-// Modern Server Administrator - Professional JavaScript Application (upgraded)
-//
-// This file contains all the original behaviour plus upgrades:
-// - Full folder uploads (via <input webkitdirectory> or multiple files upload) preserving subpaths
-// - Zip upload support (uploads .zip -> /api/upload_zip)
-// - Upload progress reporting (uses XHR so you can show progress bars)
-// - open_browser: uses backend returned URL and opens it on the client (so phones open the correct host IP)
-// - Auto-refresh on first load and on page reload (UI fetches servers immediately)
-// - Logs auto-refresh while the logs modal is open (polls /api/server_logs every 2s)
-// - Safe fallbacks and feature detection for browsers that don't support folder picks
-// - No features removed; all previous routes / behaviors preserved
-//
-// NOTE: this expects the backend endpoints we built previously:
-//   /api/servers, /api/add_server, /api/stop_server, /api/start_server, /api/delete_server,
-//   /api/server_logs/:name, /api/open_browser/:name, /api/upload_zip, /api/upload_folder, /api/system_info,
-//   /api/check_php, /api/check_node, /api/set_folder, /api/install_php
-//
-// If your HTML doesn't provide some optional UI elements (upload buttons, progress bar), the code will still work — it just won't show progress.
+// Unified Server Administrator - Professional JavaScript Application
 
-class ServerAdmin {
+class UnifiedServerAdmin {
     constructor() {
         this.currentServer = null;
         this.autoRefreshInterval = null;
         this.logsInterval = null;
-        this.selectedFiles = [];     // FileList or array used for folder upload
-        this.folderMode = false;     // whether the last selection was via folder input
         this.init();
     }
 
     init() {
         this.bindEvents();
-        // Immediately load servers (auto refresh on app entry)
         this.loadServers();
-        // Start auto-refresh (periodic)
         this.startAutoRefresh();
         this.checkSystemRequirements();
     }
 
     bindEvents() {
-        // Keep original buttons (if they exist)
-        const browseBtn = document.getElementById('browseFolderBtn');
-        if (browseBtn) browseBtn.addEventListener('click', () => this.browseFolder());
-
-        const manualBtn = document.getElementById('manualFolderBtn');
-        if (manualBtn) manualBtn.addEventListener('click', () => this.manualFolder());
-
-        const addBtn = document.getElementById('addServerBtn');
-        if (addBtn) addBtn.addEventListener('click', () => this.addServer());
-
-        const refreshBtn = document.getElementById('refreshAllBtn');
-        if (refreshBtn) refreshBtn.addEventListener('click', () => this.refreshAll());
-
-        const sysBtn = document.getElementById('systemInfoBtn');
-        if (sysBtn) sysBtn.addEventListener('click', () => this.showSystemInfo());
-
-        // Folder input (supports webkitdirectory)
-        const folderInput = document.getElementById('folderInput');
-        if (folderInput) {
-            // Accept directories (in supporting browsers)
-            // <input id="folderInput" type="file" webkitdirectory multiple />
-            folderInput.addEventListener('change', (e) => this.handleFolderSelection(e));
+        // Form controls
+        const createServerBtn = document.getElementById('createServerBtn');
+        if (createServerBtn) {
+            createServerBtn.addEventListener('click', () => this.createServer());
         }
 
-        // Upload controls (new)
-        const uploadFolderBtn = document.getElementById('uploadFolderBtn');
-        if (uploadFolderBtn) uploadFolderBtn.addEventListener('click', () => this.uploadSelectedFolder());
-
-        const uploadZipBtn = document.getElementById('uploadZipBtn');
-        if (uploadZipBtn) uploadZipBtn.addEventListener('click', () => this.uploadZipFile());
-
-        const zipInput = document.getElementById('zipInput');
-        if (zipInput) {
-            zipInput.addEventListener('change', (e) => {
-                // store selected zip for upload
-                this.zipFile = e.target.files && e.target.files.length ? e.target.files[0] : null;
-            });
+        const refreshAllBtn = document.getElementById('refreshAllBtn');
+        if (refreshAllBtn) {
+            refreshAllBtn.addEventListener('click', () => this.refreshAll());
         }
 
-        // Modal close buttons
+        const browsePathBtn = document.getElementById('browsePathBtn');
+        if (browsePathBtn) {
+            browsePathBtn.addEventListener('click', () => this.browsePath());
+        }
+
+        // System info and settings
+        const systemInfoBtn = document.getElementById('systemInfoBtn');
+        if (systemInfoBtn) {
+            systemInfoBtn.addEventListener('click', () => this.showSystemInfo());
+        }
+
+        const settingsBtn = document.getElementById('settingsBtn');
+        if (settingsBtn) {
+            settingsBtn.addEventListener('click', () => this.showSettings());
+        }
+
+        // Modal controls
         document.querySelectorAll('.close-modal').forEach(btn => {
             btn.addEventListener('click', (e) => this.closeModal(e.target.closest('.modal')));
         });
 
         // Log modal events
         const refreshLogsBtn = document.getElementById('refreshLogsBtn');
-        if (refreshLogsBtn) refreshLogsBtn.addEventListener('click', () => this.refreshLogs());
+        if (refreshLogsBtn) {
+            refreshLogsBtn.addEventListener('click', () => this.refreshLogs());
+        }
 
         const copyLogsBtn = document.getElementById('copyLogsBtn');
-        if (copyLogsBtn) copyLogsBtn.addEventListener('click', () => this.copyLogs());
-
-        const openBrowserBtn = document.getElementById('openBrowserBtn');
-        if (openBrowserBtn) openBrowserBtn.addEventListener('click', () => this.openBrowser());
-
-        // File selection: "Select all" checkbox
-        const selectAll = document.getElementById('selectAllFiles');
-        if (selectAll) selectAll.addEventListener('change', (e) => this.toggleSelectAll(e.target.checked));
+        if (copyLogsBtn) {
+            copyLogsBtn.addEventListener('click', () => this.copyLogs());
+        }
 
         // Click outside modal to close
         window.addEventListener('click', (e) => {
-            if (e.target.classList && e.target.classList.contains && e.target.classList.contains('modal')) {
+            if (e.target.classList && e.target.classList.contains('modal')) {
                 this.closeModal(e.target);
             }
         });
@@ -109,298 +73,106 @@ class ServerAdmin {
 
         // Port validation
         const portInput = document.getElementById('serverPort');
-        if (portInput) portInput.addEventListener('input', (ev) => this.validatePort(ev.target));
-
-        // Open browser action: handled on each server card's button in renderServers()
-    }
-
-    /****************
-     * Folder handling
-     ****************/
-    async browseFolder() {
-        const folderInput = document.getElementById('folderInput');
-        if (folderInput && folderInput.webkitdirectory !== undefined) {
-            // show native picker which allows folder selection in supporting browsers
-            folderInput.click();
-        } else {
-            // fallback: ask for manual path
-            this.manualFolder();
+        if (portInput) {
+            portInput.addEventListener('input', (ev) => this.validatePort(ev.target));
         }
     }
 
-    async manualFolder() {
-        const folderPath = prompt('Enter the full path to your website folder:\n\nExamples:\n- C:\\Users\\YourName\\Documents\\MyWebsite\n- /home/username/MyWebsite\n- ./my-website');
-        if (folderPath && folderPath.trim()) {
-            document.getElementById('folderPath').value = folderPath.trim();
-            // keep selectedFiles empty (we are pointing to server-side path)
-            this.selectedFiles = [];
-            this.folderMode = false;
-            await this.setFolder(folderPath.trim());
-        }
-    }
-
-    handleFolderSelection(event) {
-        // When user picks a folder using input[webkitdirectory], we get a FileList with webkitRelativePath for each file
-        const files = Array.from(event.target.files || []);
-        if (!files.length) return;
-
-        // Store files for upload (we will preserve webkitRelativePath when uploading)
-        this.selectedFiles = files;
-        this.folderMode = true;
-
-        // Derive a friendly folder name to show in the UI (the first file's webkitRelativePath's top folder)
-        let folderName = '';
-        if (files[0].webkitRelativePath) {
-            folderName = files[0].webkitRelativePath.split('/')[0];
-        } else {
-            // fallback - use file input name
-            folderName = files[0].name;
-        }
-        document.getElementById('folderPath').value = folderName;
-        // For convenience, also set the folder as current folder on the server side (optional)
-        // but many times the browser doesn't know full absolute path, so we only set folder path in field for server to import if user clicks Add Server.
-    }
-
-    /****************
-     * Upload (folder & zip)
-     ****************/
-    uploadSelectedFolder() {
-        // Called by UI "Upload Folder" button: sends selectedFiles[] with webkitRelativePath preserved
-        if (!this.selectedFiles || this.selectedFiles.length === 0) {
-            this.showNotification('No folder selected to upload. Use "Browse Folder" first.', 'error');
-            return;
-        }
-
-        const uploadBtn = document.getElementById('uploadFolderBtn');
-        const origHTML = uploadBtn ? uploadBtn.innerHTML : null;
-        if (uploadBtn) {
-            uploadBtn.innerHTML = '<i class="fas fa-spinner fa-spin"></i> Uploading...';
-            uploadBtn.disabled = true;
-        }
-
-        // Build FormData: append each file as files[] and set filename to webkitRelativePath to preserve paths
-        const fd = new FormData();
-        this.selectedFiles.forEach(f => {
-            // Preserve directory structure by using webkitRelativePath when available
-            const filename = f.webkitRelativePath ? f.webkitRelativePath : f.name;
-            fd.append('files[]', f, filename);
-        });
-        fd.append('folder_name', document.getElementById('folderPath').value || `site-${Date.now()}`);
-        // call /api/upload_folder via XHR to get upload progress
-        const xhr = new XMLHttpRequest();
-        xhr.open('POST', '/api/upload_folder', true);
-
-        xhr.upload.onprogress = (ev) => {
-            const progressBar = document.getElementById('uploadProgress');
-            if (progressBar && ev.lengthComputable) {
-                const percent = Math.round((ev.loaded / ev.total) * 100);
-                progressBar.style.width = percent + '%';
-                progressBar.textContent = `${percent}%`;
-            }
-        };
-
-        xhr.onload = async () => {
-            if (uploadBtn) { uploadBtn.innerHTML = origHTML; uploadBtn.disabled = false; }
-            const progressBar = document.getElementById('uploadProgress');
-            if (progressBar) { progressBar.style.width = '0%'; progressBar.textContent = ''; }
-            try {
-                const res = JSON.parse(xhr.responseText);
-                if (res.success) {
-                    this.showNotification('Folder uploaded successfully', 'success');
-                    // auto-register happened server-side; refresh servers list
-                    await this.loadServers();
-                } else {
-                    this.showNotification(res.message || 'Upload failed', 'error');
-                }
-            } catch (err) {
-                this.showNotification('Upload failed: invalid response', 'error');
-            }
-        };
-
-        xhr.onerror = () => {
-            if (uploadBtn) { uploadBtn.innerHTML = origHTML; uploadBtn.disabled = false; }
-            this.showNotification('Upload failed due to network error', 'error');
-        };
-
-        xhr.send(fd);
-    }
-
-    uploadZipFile() {
-        // Uploads a zip file via /api/upload_zip using XHR to allow progress
-        const zipInput = document.getElementById('zipInput');
-        let file = null;
-        if (zipInput && zipInput.files && zipInput.files.length > 0) {
-            file = zipInput.files[0];
-        } else if (this.zipFile) {
-            file = this.zipFile;
-        }
-        if (!file) {
-            this.showNotification('No zip file selected', 'error');
-            return;
-        }
-        if (!file.name.toLowerCase().endsWith('.zip')) {
-            this.showNotification('Selected file is not a .zip', 'error');
-            return;
-        }
-
-        const uploadBtn = document.getElementById('uploadZipBtn');
-        const origHtml = uploadBtn ? uploadBtn.innerHTML : null;
-        if (uploadBtn) { uploadBtn.innerHTML = '<i class="fas fa-spinner fa-spin"></i> Uploading...'; uploadBtn.disabled = true; }
-
-        const fd = new FormData();
-        fd.append('file', file, file.name);
-        fd.append('auto_register', 'true');
-
-        const xhr = new XMLHttpRequest();
-        xhr.open('POST', '/api/upload_zip', true);
-
-        xhr.upload.onprogress = (ev) => {
-            const progressBar = document.getElementById('uploadProgress');
-            if (progressBar && ev.lengthComputable) {
-                const percent = Math.round((ev.loaded / ev.total) * 100);
-                progressBar.style.width = percent + '%';
-                progressBar.textContent = `${percent}%`;
-            }
-        };
-        xhr.onload = async () => {
-            if (uploadBtn) { uploadBtn.innerHTML = origHtml; uploadBtn.disabled = false; }
-            const progressBar = document.getElementById('uploadProgress');
-            if (progressBar) { progressBar.style.width = '0%'; progressBar.textContent = ''; }
-            try {
-                const res = JSON.parse(xhr.responseText);
-                if (res.success) {
-                    this.showNotification('Zip uploaded & extracted successfully', 'success');
-                    await this.loadServers();
-                } else {
-                    this.showNotification(res.message || 'Zip extraction failed', 'error');
-                }
-            } catch (err) {
-                this.showNotification('Upload failed: invalid response', 'error');
-            }
-        };
-        xhr.onerror = () => {
-            if (uploadBtn) { uploadBtn.innerHTML = origHtml; uploadBtn.disabled = false; }
-            this.showNotification('Upload failed due to network error', 'error');
-        };
-        xhr.send(fd);
-    }
-
-    /****************
-     * Server control API calls
-     ****************/
-    async addServer() {
-        // preserved original behavior but account for folderMode (if user selected local folder files we uploaded earlier)
-        const folder = document.getElementById('folderPath').value.trim();
-        const port = parseInt(document.getElementById('serverPort').value);
+    // Server Management
+    async createServer() {
+        const name = document.getElementById('serverName').value.trim();
         const serverType = document.getElementById('serverType').value;
+        const port = parseInt(document.getElementById('serverPort').value);
+        const sitePath = document.getElementById('sitePath').value.trim();
+        const domain = document.getElementById('domain').value.trim() || null;
 
-        if (!folder) {
-            this.showNotification('Please select a website folder', 'error');
+        if (!name) {
+            this.showNotification('Server name is required', 'error');
             return;
         }
+
+        if (!sitePath) {
+            this.showNotification('Site path is required', 'error');
+            return;
+        }
+
         if (!port || port < 1000 || port > 65535) {
-            this.showNotification('Please enter a valid port number (1000-65535)', 'error');
+            this.showNotification('Port must be between 1000 and 65535', 'error');
             return;
         }
 
-        // Provide user feedback
-        const addBtn = document.getElementById('addServerBtn');
-        const originalText = addBtn ? addBtn.innerHTML : null;
-        if (addBtn) {
-            addBtn.innerHTML = '<i class="fas fa-spinner fa-spin"></i> Adding Server...';
-            addBtn.disabled = true;
+        const createBtn = document.getElementById('createServerBtn');
+        const originalText = createBtn ? createBtn.innerHTML : null;
+        
+        if (createBtn) {
+            createBtn.innerHTML = '<i class="fas fa-spinner fa-spin"></i> Creating...';
+            createBtn.disabled = true;
         }
 
         try {
-            // If folderMode (we have selected files), it may be preferable to upload those files first to backend and get a site folder.
-            // But since backend's /api/add_server will import a server-side path, we rely on the user uploading first (via Upload Folder / Zip).
-            // Here we simply post the folder path (which for remote phones might be a phrase used by backend to find the proper site in sites/).
             const response = await fetch('/api/add_server', {
                 method: 'POST',
                 headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({ folder, port, type: serverType })
+                body: JSON.stringify({
+                    name,
+                    type: serverType,
+                    port,
+                    site_path: sitePath,
+                    domain
+                })
             });
+
             const result = await response.json();
+            
             if (result.success) {
                 this.showNotification(result.message, 'success');
+                this.clearForm();
                 await this.loadServers();
             } else {
-                this.showNotification(result.message || 'Failed to add server', 'error');
+                this.showNotification(result.message || 'Failed to create server', 'error');
             }
         } catch (error) {
-            this.showNotification('Error adding server: ' + (error.message || error), 'error');
+            this.showNotification('Error creating server: ' + (error.message || error), 'error');
         } finally {
-            if (addBtn) { addBtn.innerHTML = originalText; addBtn.disabled = false; }
-        }
-    }
-
-    async stopServer(serverName) {
-        if (!confirm(`Are you sure you want to stop ${serverName}?`)) return;
-        try {
-            const res = await fetch('/api/stop_server', {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({ name: serverName })
-            });
-            const r = await res.json();
-            if (r.success) {
-                this.showNotification(r.message, 'success');
-                this.loadServers();
-            } else {
-                this.showNotification(r.message, 'error');
+            if (createBtn) {
+                createBtn.innerHTML = originalText;
+                createBtn.disabled = false;
             }
-        } catch (e) {
-            this.showNotification('Error stopping server: ' + e.message, 'error');
-        }
-    }
-
-    async startServer(serverName) {
-        try {
-            const res = await fetch('/api/start_server', {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({ name: serverName })
-            });
-            const r = await res.json();
-            if (r.success) {
-                this.showNotification(r.message, 'success');
-                this.loadServers();
-            } else {
-                this.showNotification(r.message, 'error');
-            }
-        } catch (e) {
-            this.showNotification('Error starting server: ' + e.message, 'error');
         }
     }
 
     async deleteServer(serverName) {
-        if (!confirm(`Are you sure you want to delete server "${serverName}"? This action cannot be undone.`)) return;
+        if (!confirm(`Are you sure you want to delete server "${serverName}"? This action cannot be undone.`)) {
+            return;
+        }
+
         try {
-            const res = await fetch('/api/delete_server', {
+            const response = await fetch('/api/delete_server', {
                 method: 'POST',
                 headers: { 'Content-Type': 'application/json' },
                 body: JSON.stringify({ name: serverName })
             });
-            const r = await res.json();
-            if (r.success) {
-                this.showNotification(r.message, 'success');
-                this.loadServers();
+
+            const result = await response.json();
+            
+            if (result.success) {
+                this.showNotification(result.message, 'success');
+                await this.loadServers();
             } else {
-                this.showNotification(r.message || 'Failed to delete', 'error');
+                this.showNotification(result.message || 'Failed to delete server', 'error');
             }
-        } catch (e) {
-            this.showNotification('Error deleting server: ' + e.message, 'error');
+        } catch (error) {
+            this.showNotification('Error deleting server: ' + (error.message || error), 'error');
         }
     }
 
-    /****************
-     * Server listing + rendering
-     ****************/
+    // Server Listing
     async loadServers() {
         try {
             const response = await fetch('/api/servers');
             const servers = await response.json();
             this.renderServers(servers);
+            this.updateStats(servers);
         } catch (error) {
             console.error('Error loading servers:', error);
             this.showNotification('Error loading servers', 'error');
@@ -415,22 +187,18 @@ class ServerAdmin {
             container.innerHTML = `
                 <div class="empty-state">
                     <i class="fas fa-server"></i>
-                    <h3 style="color: var(--text-primary);">No servers registered</h3>
-                    <p style="color: var(--text-secondary);">Upload a folder or add a server to get started</p>
+                    <h3>No servers configured</h3>
+                    <p>Create your first server to get started</p>
                 </div>
             `;
-            if (document.getElementById('serverCount')) document.getElementById('serverCount').textContent = '0 servers running';
             return;
         }
 
-        let runningCount = 0;
         Object.values(servers).forEach(server => {
-            if (server.status === 'Running') runningCount++;
-
             const serverCard = document.createElement('div');
             serverCard.className = `server-card ${server.status.toLowerCase()}`;
 
-            // header
+            // Header
             const header = document.createElement('div');
             header.className = 'server-header';
 
@@ -438,16 +206,13 @@ class ServerAdmin {
             serverInfo.className = 'server-info';
 
             const serverName = document.createElement('h3');
-            serverName.style.cssText = 'color: var(--text-primary); font-weight: 600;';
             serverName.textContent = server.name;
 
-            const serverFolder = document.createElement('p');
-            serverFolder.style.cssText = 'color: var(--text-secondary); font-weight: 500;';
-            // show folder name only
-            serverFolder.textContent = (server.folder || '').split('/').pop();
+            const serverType = document.createElement('p');
+            serverType.textContent = `${server.type} Server`;
 
             serverInfo.appendChild(serverName);
-            serverInfo.appendChild(serverFolder);
+            serverInfo.appendChild(serverType);
 
             const serverStatus = document.createElement('div');
             serverStatus.className = 'server-status';
@@ -460,27 +225,35 @@ class ServerAdmin {
             header.appendChild(serverInfo);
             header.appendChild(serverStatus);
 
-            // details
+            // Details
             const details = document.createElement('div');
             details.className = 'server-details';
+            
             const detailsData = [
                 { label: 'Port', value: server.port || '—' },
-                { label: 'Type', value: server.type || 'HTTP' },
+                { label: 'Type', value: server.type || 'Static' },
                 { label: 'Started', value: server.start_time || '—' },
-                // Do not assume localhost: show a helpful hint instead (client should use open button to get correct URL)
-                { label: 'URL', value: server.port ? `Click "Open Browser" to get device URL` : 'Not started' }
+                { label: 'Domain', value: server.domain || '—' }
             ];
+
             detailsData.forEach(item => {
                 const detailItem = document.createElement('div');
                 detailItem.className = 'detail-item';
-                const label = document.createElement('div'); label.className = 'detail-label'; label.textContent = item.label;
-                const value = document.createElement('div'); value.className = 'detail-value'; value.style.cssText = 'color: var(--text-primary); font-weight: 600;';
+                
+                const label = document.createElement('div');
+                label.className = 'detail-label';
+                label.textContent = item.label;
+                
+                const value = document.createElement('div');
+                value.className = 'detail-value';
                 value.textContent = item.value;
-                detailItem.appendChild(label); detailItem.appendChild(value);
+                
+                detailItem.appendChild(label);
+                detailItem.appendChild(value);
                 details.appendChild(detailItem);
             });
 
-            // actions
+            // Actions
             const actions = document.createElement('div');
             actions.className = 'server-actions';
 
@@ -490,25 +263,6 @@ class ServerAdmin {
             viewLogsBtn.innerHTML = '<i class="fas fa-file-alt"></i> View Logs';
             viewLogsBtn.onclick = () => this.showLogs(server.name);
 
-            const openBrowserBtn = document.createElement('button');
-            openBrowserBtn.className = 'action-btn open-browser';
-            openBrowserBtn.title = 'Open in Browser';
-            openBrowserBtn.innerHTML = '<i class="fas fa-external-link-alt"></i> Open Browser';
-            openBrowserBtn.onclick = () => this.openServerInBrowser(server.name);
-
-            const startStopBtn = document.createElement('button');
-            if (server.status === 'Running') {
-                startStopBtn.className = 'action-btn stop-server';
-                startStopBtn.title = 'Stop Server';
-                startStopBtn.innerHTML = '<i class="fas fa-stop"></i> Stop Server';
-                startStopBtn.onclick = () => this.stopServer(server.name);
-            } else {
-                startStopBtn.className = 'action-btn start-server';
-                startStopBtn.title = 'Start Server';
-                startStopBtn.innerHTML = '<i class="fas fa-play"></i> Start Server';
-                startStopBtn.onclick = () => this.startServer(server.name);
-            }
-
             const deleteBtn = document.createElement('button');
             deleteBtn.className = 'action-btn delete-server';
             deleteBtn.title = 'Delete Server';
@@ -516,8 +270,6 @@ class ServerAdmin {
             deleteBtn.onclick = () => this.deleteServer(server.name);
 
             actions.appendChild(viewLogsBtn);
-            actions.appendChild(openBrowserBtn);
-            actions.appendChild(startStopBtn);
             actions.appendChild(deleteBtn);
 
             serverCard.appendChild(header);
@@ -525,56 +277,95 @@ class ServerAdmin {
             serverCard.appendChild(actions);
             container.appendChild(serverCard);
         });
-
-        if (document.getElementById('serverCount')) document.getElementById('serverCount').textContent = `${runningCount} servers running`;
     }
 
-    /****************
-     * Logs UI
-     ****************/
+    updateStats(servers) {
+        const totalServers = Object.keys(servers).length;
+        const runningServers = Object.values(servers).filter(s => s.status === 'Running').length;
+        const phpServers = Object.values(servers).filter(s => s.type === 'PHP').length;
+        const nodeServers = Object.values(servers).filter(s => s.type === 'Node.js').length;
+
+        const totalElement = document.getElementById('totalServers');
+        const runningElement = document.getElementById('runningServers');
+        const phpElement = document.getElementById('phpServers');
+        const nodeElement = document.getElementById('nodeServers');
+
+        if (totalElement) totalElement.textContent = totalServers;
+        if (runningElement) runningElement.textContent = runningServers;
+        if (phpElement) phpElement.textContent = phpServers;
+        if (nodeElement) nodeElement.textContent = nodeServers;
+
+        const serverCountElement = document.getElementById('serverCount');
+        if (serverCountElement) {
+            serverCountElement.textContent = `${totalServers} server${totalServers !== 1 ? 's' : ''}`;
+        }
+    }
+
+    // Logs Management
     async showLogs(serverName) {
         this.currentServer = serverName;
         const title = document.getElementById('logModalTitle');
-        if (title) title.innerHTML = `<i class="fas fa-file-alt"></i> Logs - ${serverName}`;
+        if (title) {
+            title.innerHTML = `<i class="fas fa-file-alt"></i> Logs - ${serverName}`;
+        }
+        
         const modal = document.getElementById('logModal');
-        if (modal) modal.classList.add('show');
-        // Immediately load logs and start polling
+        if (modal) {
+            modal.classList.add('show');
+        }
+        
         await this.refreshLogs();
         this.startLogsAutoRefresh();
     }
 
     async refreshLogs() {
         if (!this.currentServer) return;
+        
         try {
             const response = await fetch(`/api/server_logs/${encodeURIComponent(this.currentServer)}`);
             const data = await response.json();
             const logContent = document.getElementById('logContent');
+            
             if (!logContent) return;
+            
             logContent.innerHTML = '';
+            
             if (!data.logs || data.logs.length === 0) {
                 logContent.textContent = 'No logs available yet...';
                 return;
             }
+            
             data.logs.forEach(log => {
                 const logLine = document.createElement('div');
                 const timestamp = log.timestamp || new Date().toLocaleTimeString();
                 const message = log.message || '';
                 const type = log.type || 'info';
-                logLine.innerHTML = `<span style="color: var(--text-muted);">[${timestamp}]</span> <span style="color:${type === 'error' ? 'var(--error)' : 'var(--text-primary)'};">${this.escapeHtml(message)}</span>`;
+                
+                logLine.innerHTML = `
+                    <span style="color: var(--text-muted);">[${timestamp}]</span> 
+                    <span style="color: ${type === 'error' ? 'var(--error)' : 'var(--text-primary)'};">${this.escapeHtml(message)}</span>
+                `;
                 logContent.appendChild(logLine);
             });
+            
             logContent.scrollTop = logContent.scrollHeight;
         } catch (error) {
             console.error('Error loading logs:', error);
             const logContent = document.getElementById('logContent');
-            if (logContent) logContent.textContent = 'Error loading logs: ' + (error.message || error);
+            if (logContent) {
+                logContent.textContent = 'Error loading logs: ' + (error.message || error);
+            }
         }
     }
 
     startLogsAutoRefresh() {
-        // stop existing
         this.stopLogsAutoRefresh();
         this.logsInterval = setInterval(() => this.refreshLogs(), 2000);
+        
+        const liveBadge = document.getElementById('logLiveBadge');
+        if (liveBadge) {
+            liveBadge.style.display = 'flex';
+        }
     }
 
     stopLogsAutoRefresh() {
@@ -582,11 +373,20 @@ class ServerAdmin {
             clearInterval(this.logsInterval);
             this.logsInterval = null;
         }
+        
+        const liveBadge = document.getElementById('logLiveBadge');
+        if (liveBadge) {
+            liveBadge.style.display = 'none';
+        }
     }
 
     async copyLogs() {
         const logContent = document.getElementById('logContent');
-        if (!logContent) return this.showNotification('No logs to copy', 'warning');
+        if (!logContent) {
+            this.showNotification('No logs to copy', 'warning');
+            return;
+        }
+        
         try {
             await navigator.clipboard.writeText(logContent.innerText);
             this.showNotification('Logs copied to clipboard', 'success');
@@ -595,119 +395,98 @@ class ServerAdmin {
         }
     }
 
-    /****************
-     * Open Browser behavior (uses backend url)
-     ****************/
-    async openServerInBrowser(serverName) {
-        try {
-            const response = await fetch(`/api/open_browser/${encodeURIComponent(serverName)}`);
-            const result = await response.json();
-            // backend returns { success, url, message }
-            if (result.success && result.url) {
-                // open on client device automatically
-                window.open(result.url, '_blank');
-                this.showNotification(`Opened ${result.url}`, 'success');
-            } else if (result.success && !result.url) {
-                // fallback message
-                this.showNotification(result.message || 'URL returned empty', 'warning');
-            } else {
-                // error from backend
-                this.showNotification(result.message || 'Failed to open', 'error');
-            }
-        } catch (error) {
-            this.showNotification('Error opening browser: ' + (error.message || error), 'error');
-        }
-    }
-
-    async openBrowser() {
-        if (!this.currentServer) return;
-        await this.openServerInBrowser(this.currentServer);
-    }
-
-    /****************
-     * Misc utilities / small UI bits
-     ****************/
-    async refreshAll() {
-        await this.loadServers();
-        this.updateStatus('Refreshed', 'success');
-    }
-
+    // System Information
     async showSystemInfo() {
         try {
             const response = await fetch('/api/system_info');
             const info = await response.json();
             const content = document.getElementById('systemInfoContent');
+            
             if (!content) return;
+            
             content.innerHTML = `
                 <h4>System Information</h4>
-                <p><strong>Operating System:</strong> ${this.escapeHtml(info.os || '')}</p>
-                <p><strong>Architecture:</strong> ${this.escapeHtml(info.architecture || '')}</p>
-                <p><strong>Python Version:</strong> ${this.escapeHtml(info.python_version || '')}</p>
-                <p><strong>CPU Cores:</strong> ${this.escapeHtml(info.cpu_cores || '')}</p>
-                <p><strong>Memory Total (GB):</strong> ${this.escapeHtml(String(info.memory_total_gb || info.memory_total || ''))}</p>
-                <p><strong>Memory Available (GB):</strong> ${this.escapeHtml(String(info.memory_available_gb || info.memory_available || ''))}</p>
-                <p><strong>Disk Free (GB):</strong> ${this.escapeHtml(String(info.disk_free_gb || info.disk_free || ''))}</p>
+                <p><strong>Operating System:</strong> ${this.escapeHtml(info.os || 'Unknown')}</p>
+                <p><strong>CPU Cores:</strong> ${this.escapeHtml(info.cpu_cores || 'Unknown')}</p>
+                <p><strong>Python Version:</strong> ${this.escapeHtml(info.python_version || 'Unknown')}</p>
                 <p><strong>Active Servers:</strong> ${this.escapeHtml(String(info.active_servers || '0'))}</p>
-                <p><strong>PHP Available:</strong> ${(info.php_available ? 'Yes' : 'No')}</p>
-                ${info.php_version ? `<p><strong>PHP Version:</strong> ${this.escapeHtml(info.php_version)}</p>` : ''}
-                <p><strong>Node.js Available:</strong> ${(info.node_available ? 'Yes' : 'No')}</p>
-                ${info.node_version ? `<p><strong>Node.js Version:</strong> ${this.escapeHtml(info.node_version)}</p>` : ''}
-                <p><strong>Sites Folder:</strong> ${this.escapeHtml(info.sites_folder || info.current_folder || '')}</p>
+                <p><strong>Sites Folder:</strong> ${this.escapeHtml(info.sites_folder || 'Unknown')}</p>
+                <p><strong>PHP Available:</strong> ${info.php_available ? 'Yes' : 'No'}</p>
+                <p><strong>Node.js Available:</strong> ${info.node_available ? 'Yes' : 'No'}</p>
+                <p><strong>Nginx Available:</strong> ${info.nginx_available ? 'Yes' : 'No'}</p>
             `;
+            
             document.getElementById('systemModal').classList.add('show');
         } catch (error) {
             this.showNotification('Error loading system info: ' + (error.message || error), 'error');
         }
     }
 
+    showSettings() {
+        document.getElementById('settingsModal').classList.add('show');
+    }
+
+    // Utility Functions
+    async refreshAll() {
+        await this.loadServers();
+        this.updateStatus('Refreshed', 'success');
+    }
+
     async checkSystemRequirements() {
         try {
-            const [phpResponse, nodeResponse] = await Promise.all([fetch('/api/check_php'), fetch('/api/check_node')]);
-            const phpInfo = await phpResponse.json();
-            const nodeInfo = await nodeResponse.json();
-            if (!phpInfo.available) {
-                this.showNotification('PHP is not available - PHP servers will not work. Use the Install PHP button.', 'warning');
-                const btn = document.getElementById('installPhpBtn'); if (btn) btn.style.display = 'inline-flex';
-            } else {
-                this.showNotification(`PHP detected: ${phpInfo.version}`, 'success');
+            const response = await fetch('/api/system_info');
+            const info = await response.json();
+            
+            if (!info.nginx_available) {
+                this.showNotification('Nginx is not available - server creation will not work', 'error');
             }
-            if (!nodeInfo.available) {
+            
+            if (!info.php_available) {
+                this.showNotification('PHP is not available - PHP servers will not work', 'warning');
+            }
+            
+            if (!info.node_available) {
                 this.showNotification('Node.js is not available - Node.js servers will not work', 'warning');
-            } else {
-                this.showNotification(`Node.js detected: ${nodeInfo.version}`, 'success');
             }
         } catch (error) {
             console.error('Error checking system requirements:', error);
         }
     }
 
-    async installPHP() {
-        const installBtn = document.getElementById('installPhpBtn');
-        if (!installBtn) return this.showNotification('Install button missing', 'error');
-        const orig = installBtn.innerHTML;
-        installBtn.innerHTML = '<i class="fas fa-spinner fa-spin"></i> Installing PHP...';
-        installBtn.disabled = true;
-        try {
-            const response = await fetch('/api/install_php', { method: 'POST' });
-            const res = await response.json();
-            if (res.success) {
-                this.showNotification(res.message, 'success');
-                installBtn.style.display = 'none';
-            } else {
-                this.showNotification(res.message || 'Install failed', 'error');
-            }
-        } catch (e) {
-            this.showNotification('Error installing PHP: ' + (e.message || e), 'error');
-        } finally {
-            installBtn.innerHTML = orig;
-            installBtn.disabled = false;
+    browsePath() {
+        // For now, just show a prompt. In a real implementation, you might use a file picker
+        const path = prompt('Enter the full path to your website folder:');
+        if (path && path.trim()) {
+            document.getElementById('sitePath').value = path.trim();
         }
     }
 
+    clearForm() {
+        document.getElementById('serverName').value = '';
+        document.getElementById('serverType').value = 'Static';
+        document.getElementById('serverPort').value = '8000';
+        document.getElementById('sitePath').value = '';
+        document.getElementById('domain').value = '';
+    }
+
+    validatePort(input) {
+        const port = parseInt(input.value);
+        if (port < 1000 || port > 65535) {
+            input.style.borderColor = 'var(--error)';
+        } else {
+            input.style.borderColor = 'var(--border-primary)';
+        }
+    }
+
+    // Modal Management
     closeModal(modal) {
-        if (modal) modal.classList.remove('show');
-        // stop any log polling when closing logs modal
-        if (modal && modal.id === 'logModal') this.stopLogsAutoRefresh();
+        if (modal) {
+            modal.classList.remove('show');
+        }
+        
+        if (modal && modal.id === 'logModal') {
+            this.stopLogsAutoRefresh();
+        }
     }
 
     closeAllModals() {
@@ -718,15 +497,19 @@ class ServerAdmin {
     updateStatus(text, type) {
         const statusDot = document.getElementById('statusDot');
         const statusText = document.getElementById('statusText');
+        
         if (!statusText) return;
+        
         statusText.textContent = text;
+        
         if (statusDot) {
             if (type === 'success') statusDot.style.background = 'var(--success)';
             else if (type === 'error') statusDot.style.background = 'var(--error)';
             else if (type === 'warning') statusDot.style.background = 'var(--warning)';
             else statusDot.style.background = 'var(--info)';
         }
-        // reset
+        
+        // Reset after 3 seconds
         setTimeout(() => {
             if (statusText) statusText.textContent = 'Ready';
             if (statusDot) statusDot.style.background = 'var(--success)';
@@ -736,41 +519,41 @@ class ServerAdmin {
     showNotification(message, type = 'info') {
         const container = document.getElementById('notificationContainer');
         if (!container) {
-            // fallback alert
-            if (type === 'error') console.error(message);
-            else console.log(message);
+            console.log(`[${type.toUpperCase()}] ${message}`);
             return;
         }
+        
         const notification = document.createElement('div');
         notification.className = `notification ${type}`;
+        
         const icon = {
             success: 'fa-check-circle',
             error: 'fa-exclamation-circle',
             warning: 'fa-exclamation-triangle',
             info: 'fa-info-circle'
         }[type] || 'fa-info-circle';
+        
         notification.innerHTML = `<i class="fas ${icon}"></i><span>${this.escapeHtml(message)}</span>`;
         container.appendChild(notification);
+        
         setTimeout(() => {
             if (notification.parentNode) {
                 notification.style.animation = 'slideOutRight 0.3s ease';
-                setTimeout(() => { if (notification.parentNode) notification.parentNode.removeChild(notification); }, 300);
+                setTimeout(() => {
+                    if (notification.parentNode) {
+                        notification.parentNode.removeChild(notification);
+                    }
+                }, 300);
             }
         }, 5000);
     }
 
-    /****************
-     * Auto refresh + cleanup
-     ****************/
+    // Auto Refresh
     startAutoRefresh() {
-        // Immediately refresh on start (ensures UI is always up-to-date)
         this.loadServers();
-        // Use 30s poll by default like previous code (but first load is instant)
-        const interval = 30000;
-        if (this.autoRefreshInterval) clearInterval(this.autoRefreshInterval);
         this.autoRefreshInterval = setInterval(() => {
             this.loadServers();
-        }, interval);
+        }, 30000); // 30 seconds
     }
 
     stopAutoRefresh() {
@@ -788,53 +571,18 @@ class ServerAdmin {
             .replace(/"/g, "&quot;")
             .replace(/'/g, "&#039;");
     }
-
-    validatePort(input) {
-        const port = parseInt(input.value);
-        if (port < 1000 || port > 65535) {
-            input.style.borderColor = 'var(--error)';
-        } else {
-            input.style.borderColor = 'var(--border-primary)';
-        }
-    }
-
-    toggleSelectAll(checked) {
-        // This was originally intended for selecting individual file checkboxes;
-        // keep as a noop to not break existing UI unless you add file list checkboxes.
-        this.showNotification('Toggle select all not wired to file list UI (implement if needed)', 'info');
-    }
-
-    /****************
-     * Helpers for debugging / convenience
-     ****************/
-    async installPHP_if_button_exists() {
-        const btn = document.getElementById('installPhpBtn');
-        if (btn) btn.addEventListener('click', () => this.installPHP());
-    }
 }
 
-// Ensure small CSS animation exists (keeps original style addition)
-const style = document.createElement('style');
-style.textContent = `
-    @keyframes slideOutRight {
-        from { transform: translateX(0); opacity: 1; }
-        to { transform: translateX(100%); opacity: 0; }
-    }
-    /* Minimal upload progress element (if present) */
-    #uploadProgressBar { width: 100%; background: #111; height: 6px; border-radius: 3px; overflow: hidden; margin-top:6px; }
-    #uploadProgressBar > .bar { width: 0%; height:100%; text-align:center; font-size:11px; line-height:6px; color:#fff; }
-`;
-document.head.appendChild(style);
+// Initialize application
+let unifiedServerAdmin;
 
-// Initialize
-let serverAdmin;
 document.addEventListener('DOMContentLoaded', () => {
-    serverAdmin = new ServerAdmin();
-    // Hook install PHP button if exists
-    serverAdmin.installPHP_if_button_exists();
+    unifiedServerAdmin = new UnifiedServerAdmin();
 });
 
 // Cleanup on unload
 window.addEventListener('beforeunload', () => {
-    if (serverAdmin) serverAdmin.stopAutoRefresh();
+    if (unifiedServerAdmin) {
+        unifiedServerAdmin.stopAutoRefresh();
+    }
 });
