+++ web_server_admin.py
@@ -1,14 +1,20 @@
 #!/usr/bin/env python3
 """
-Modern Server Administrator - Production (final polish)
+Unified Server Administrator - Production Ready
+===============================================
 
-- Persistent server registry (servers.json)
-- Background monitor for auto-refresh (updates UI without manual refresh)
-- Fixed open_browser (returns usable URL, optional local open)
-- Robust uploads (zip, multiple-files), atomic extraction & recursive import
-- Per-server rotating logs + in-memory queues (UI tail + file fallback)
-- ThreadPoolExecutor for async IO
-- Toggleable runtime options (exposed as constants for now)
+A professional, production-ready web server administration tool that unifies:
+- PHP via php-fpm
+- Node.js via internal proxy
+- Static files via nginx
+- All managed through a single nginx reverse proxy
+
+Features:
+- GitHub-like professional dark theme UI
+- Nginx reverse proxy configuration
+- Unified server management
+- Real-time monitoring and logs
+- Professional deployment ready
 """
 
 import os
@@ -22,10 +28,11 @@ import zipfile
 import subprocess
 import threading
 import time
+import yaml
 from datetime import datetime
 from pathlib import Path
 from concurrent.futures import ThreadPoolExecutor
-from typing import Tuple
+from typing import Dict, List, Optional, Tuple
 
 import logging
 import logging.handlers
@@ -38,63 +45,96 @@ try:
 except Exception:
     PSUTIL_AVAILABLE = False
 
-from flask import Flask, render_template, request, jsonify
+from flask import Flask, render_template, request, jsonify, send_from_directory
 from werkzeug.utils import secure_filename
 
-# -------------------------
-# Runtime options (expose these)
-# -------------------------
-RUNTIME_OPTIONS = {
-    'AUTO_REFRESH_ON_LOAD': True,      # backend ensures fresh server states when dashboard loads
-    'AUTO_REFRESH_INTERVAL': 2.0,      # seconds between background refreshes (low = faster UI updates)
-    'AUTO_RESTORE_RUNNING': False,     # attempt to restart servers that were running at shutdown
-    'AUTO_START_DISCOVERED': False,    # whether to auto-start discovered sites in sites/
-    'BIND_ALL_DEFAULT': True,          # child servers bind to 0.0.0.0 by default for multi-device access
-    'MAX_WORKERS': 32,                 # threadpool size for concurrent uploads/etc.
-    'LOG_TAIL_LINES': 200,             # fallback lines read from file if in-memory queue empty
-    'DELETE_SITE_ON_REMOVE': False,    # default behavior when deleting a server
-    'ALLOW_REMOTE_OPEN_BROWSER': True  # allow returning remote host IP in open_browser response
-}
+# =========================
+# Configuration
+# =========================
+class Config:
+    # Server Configuration
+    HOST = '0.0.0.0'
+    PORT = 5000
+    DEBUG = False
+    
+    # Paths
+    ROOT = Path.cwd()
+    UPLOAD_FOLDER = ROOT / 'uploads'
+    SITES_FOLDER = ROOT / 'sites'
+    LOGS_FOLDER = ROOT / 'logs'
+    NGINX_CONFIG_DIR = ROOT / 'nginx_configs'
+    PHP_FPM_CONFIG_DIR = ROOT / 'php_fpm_configs'
+    PERSIST_FILE = ROOT / 'servers.json'
+    PHP_LOCAL_FOLDER = ROOT / 'php'
+    
+    # Nginx Configuration
+    NGINX_MAIN_CONFIG = '/etc/nginx/nginx.conf'
+    NGINX_SITES_AVAILABLE = '/etc/nginx/sites-available'
+    NGINX_SITES_ENABLED = '/etc/nginx/sites-enabled'
+    
+    # PHP-FPM Configuration
+    PHP_FPM_POOL_DIR = '/etc/php/8.1/fpm/pool.d'
+    PHP_FPM_SOCKET_DIR = '/var/run/php'
+    
+    # Allowed file extensions
+    ALLOWED_EXTENSIONS = {
+        'txt', 'pdf', 'png', 'jpg', 'jpeg', 'gif', 'mp4', 'avi', 'mov', 'mkv', 'wmv', 'flv', 'webm',
+        'php', 'html', 'css', 'js', 'json', 'xml', 'md', 'py', 'sql', 'zip', 'rar', '7z'
+    }
+    
+    # Runtime options
+    AUTO_REFRESH_INTERVAL = 2.0
+    MAX_WORKERS = 32
+    LOG_TAIL_LINES = 200
+    BIND_ALL_DEFAULT = True
 
-# -------------------------
-# Paths and globals
-# -------------------------
+# =========================
+# Initialize Flask App
+# =========================
 app = Flask(__name__)
-app.secret_key = 'modern_server_admin_2024'
+app.secret_key = 'unified_server_admin_2024_production'
 
-ROOT = Path.cwd()
-UPLOAD_FOLDER = ROOT / 'uploads'
-SITES_FOLDER = ROOT / 'sites'
-LOGS_FOLDER = ROOT / 'logs'
-PERSIST_FILE = ROOT / 'servers.json'
-PHP_LOCAL_FOLDER = ROOT / 'php'
+# Create necessary directories
+for directory in [Config.UPLOAD_FOLDER, Config.SITES_FOLDER, Config.LOGS_FOLDER, 
+                 Config.NGINX_CONFIG_DIR, Config.PHP_FPM_CONFIG_DIR]:
+    os.makedirs(directory, exist_ok=True)
 
-ALLOWED_EXTENSIONS = {'txt','pdf','png','jpg','jpeg','gif','mp4','avi','mov','mkv','wmv','flv','webm',
-                      'php','html','css','js','json','xml','md','py','sql','zip','rar','7z'}
+# Thread pool for concurrent operations
+EXECUTOR = ThreadPoolExecutor(max_workers=Config.MAX_WORKERS)
 
-for d in (UPLOAD_FOLDER, SITES_FOLDER, LOGS_FOLDER, PHP_LOCAL_FOLDER):
-    os.makedirs(d, exist_ok=True)
-
-EXECUTOR = ThreadPoolExecutor(max_workers=RUNTIME_OPTIONS['MAX_WORKERS'])
-
-# in-memory state
+# Global state
 servers = {}           # name -> metadata (persisted)
 server_processes = {}  # name -> subprocess.Popen
 log_queues = {}        # name -> Queue()
+nginx_configs = {}     # name -> nginx config path
 
 LOCK = threading.RLock()
 
-# Logging
-app_logger = logging.getLogger('msa_app')
-app_logger.setLevel(logging.DEBUG)
-fh = logging.handlers.RotatingFileHandler(str(LOGS_FOLDER / 'msa_app.log'), maxBytes=5_000_000, backupCount=5, encoding='utf-8')
-fh.setFormatter(logging.Formatter('%(asctime)s %(levelname)s: %(message)s'))
-app_logger.addHandler(fh)
-app_logger.addHandler(logging.StreamHandler(sys.stdout))
+# =========================
+# Logging Setup
+# =========================
+def setup_logging():
+    app_logger = logging.getLogger('usa_app')
+    app_logger.setLevel(logging.DEBUG)
+    
+    # File handler
+    fh = logging.handlers.RotatingFileHandler(
+        str(Config.LOGS_FOLDER / 'usa_app.log'), 
+        maxBytes=5_000_000, 
+        backupCount=5, 
+        encoding='utf-8'
+    )
+    fh.setFormatter(logging.Formatter('%(asctime)s %(levelname)s: %(message)s'))
+    app_logger.addHandler(fh)
+    app_logger.addHandler(logging.StreamHandler(sys.stdout))
+    
+    return app_logger
 
-# -------------------------
-# Utility functions
-# -------------------------
+app_logger = setup_logging()
+
+# =========================
+# Utility Functions
+# =========================
 def now_str():
     return datetime.now().strftime('%Y-%m-%d %H:%M:%S')
 
@@ -102,24 +142,23 @@ def human_size(path):
     try:
         s = os.path.getsize(path)
         for unit in ['B','KB','MB','GB','TB']:
-            if s < 1024.0: return f"{s:.1f} {unit}"
+            if s < 1024.0: 
+                return f"{s:.1f} {unit}"
             s /= 1024.0
     except:
         return "Unknown"
 
 def allowed_file(filename):
-    return '.' in filename and filename.rsplit('.',1)[1].lower() in ALLOWED_EXTENSIONS
+    return '.' in filename and filename.rsplit('.',1)[1].lower() in Config.ALLOWED_EXTENSIONS
 
 def secure_name(n):
     return secure_filename(n) or f"site-{int(time.time())}"
 
-# port check: prefer psutil for faster accurate checking; fallback to socket connect
-def is_port_in_use(port:int, host='0.0.0.0') -> bool:
+def is_port_in_use(port: int, host='0.0.0.0') -> bool:
     try:
         if PSUTIL_AVAILABLE:
             for conn in psutil.net_connections():
                 if conn.laddr and conn.laddr.port == int(port):
-                    # if host specified, check ip matches or host is wildcard
                     if host in ('0.0.0.0','') or (conn.laddr.ip == host) or conn.laddr.ip in ('0.0.0.0','127.0.0.1','::'):
                         return True
             return False
@@ -144,17 +183,20 @@ def tail_lines(filepath, n=200):
     except Exception:
         return []
 
-# -------------------------
-# System checks
-# -------------------------
+# =========================
+# System Checks
+# =========================
 def get_php_path():
-    local = str(PHP_LOCAL_FOLDER / ('php.exe' if platform.system() == 'Windows' else 'php'))
-    if Path(local).exists(): return local
+    local = str(Config.PHP_LOCAL_FOLDER / ('php.exe' if platform.system() == 'Windows' else 'php'))
+    if Path(local).exists(): 
+        return local
     w = shutil.which('php')
-    if w: return w
+    if w: 
+        return w
     if platform.system() == 'Windows':
         for p in [r'C:\php\php.exe', r'C:\xampp\php\php.exe', r'C:\wamp\bin\php\php.exe']:
-            if Path(p).exists(): return p
+            if Path(p).exists(): 
+                return p
     return 'php'
 
 def check_php_available():
@@ -176,335 +218,658 @@ def check_node_available():
     except:
         return False
 
-# -------------------------
-# Logging helpers
-# -------------------------
-def get_server_logger(name):
-    logger = logging.getLogger(f"msa_server_{name}")
-    if logger.handlers:
-        return logger
-    logger.setLevel(logging.INFO)
-    log_file = LOGS_FOLDER / f"{name}.log"
-    fh = logging.handlers.RotatingFileHandler(str(log_file), maxBytes=2_000_000, backupCount=3, encoding='utf-8')
-    fh.setFormatter(logging.Formatter('%(asctime)s %(levelname)s: %(message)s'))
-    logger.addHandler(fh)
-    return logger
+def check_nginx_available():
+    try:
+        r = subprocess.run(['nginx', '-v'], capture_output=True, text=True, timeout=3)
+        return r.returncode == 0
+    except:
+        return False
 
-# -------------------------
-# Persistence helpers
-# -------------------------
+# =========================
+# Nginx Configuration Management
+# =========================
+class NginxManager:
+    def __init__(self):
+        self.config_dir = Config.NGINX_CONFIG_DIR
+        self.sites_available = Config.NGINX_SITES_AVAILABLE
+        self.sites_enabled = Config.NGINX_SITES_ENABLED
+    
+    def generate_nginx_config(self, server_name: str, server_type: str, port: int, 
+                            site_path: str, domain: str = None) -> str:
+        """Generate nginx configuration for a server"""
+        
+        if server_type == 'PHP':
+            return self._generate_php_config(server_name, port, site_path, domain)
+        elif server_type == 'Node.js':
+            return self._generate_nodejs_config(server_name, port, site_path, domain)
+        else:  # Static files
+            return self._generate_static_config(server_name, port, site_path, domain)
+    
+    def _generate_php_config(self, server_name: str, port: int, site_path: str, domain: str = None) -> str:
+        """Generate nginx config for PHP with php-fpm"""
+        server_name_config = domain or f"{server_name}.local"
+        
+        config = f"""
+server {{
+    listen {port};
+    server_name {server_name_config};
+    root {site_path};
+    index index.php index.html index.htm;
+    
+    # Security headers
+    add_header X-Frame-Options "SAMEORIGIN" always;
+    add_header X-XSS-Protection "1; mode=block" always;
+    add_header X-Content-Type-Options "nosniff" always;
+    add_header Referrer-Policy "no-referrer-when-downgrade" always;
+    add_header Content-Security-Policy "default-src 'self' http: https: data: blob: 'unsafe-inline'" always;
+    
+    # Gzip compression
+    gzip on;
+    gzip_vary on;
+    gzip_min_length 1024;
+    gzip_proxied expired no-cache no-store private must-revalidate auth;
+    gzip_types text/plain text/css text/xml text/javascript application/x-javascript application/xml+rss;
+    
+    # PHP processing
+    location ~ \\.php$ {{
+        try_files $uri =404;
+        fastcgi_split_path_info ^(.+\\.php)(/.+)$;
+        fastcgi_pass unix:/var/run/php/php8.1-fpm-{server_name}.sock;
+        fastcgi_index index.php;
+        fastcgi_param SCRIPT_FILENAME $document_root$fastcgi_script_name;
+        include fastcgi_params;
+        
+        # Security
+        fastcgi_hide_header X-Powered-By;
+        fastcgi_read_timeout 300;
+        fastcgi_connect_timeout 300;
+        fastcgi_send_timeout 300;
+    }}
+    
+    # Static files
+    location / {{
+        try_files $uri $uri/ /index.php?$query_string;
+        
+        # Cache static files
+        location ~* \\.(jpg|jpeg|png|gif|ico|css|js|woff|woff2|ttf|svg)$ {{
+            expires 1y;
+            add_header Cache-Control "public, immutable";
+        }}
+    }}
+    
+    # Deny access to hidden files
+    location ~ /\\. {{
+        deny all;
+    }}
+    
+    # Logging
+    access_log /var/log/nginx/{server_name}_access.log;
+    error_log /var/log/nginx/{server_name}_error.log;
+}}
+"""
+        return config.strip()
+    
+    def _generate_nodejs_config(self, server_name: str, port: int, site_path: str, domain: str = None) -> str:
+        """Generate nginx config for Node.js reverse proxy"""
+        server_name_config = domain or f"{server_name}.local"
+        
+        config = f"""
+upstream {server_name}_backend {{
+    server 127.0.0.1:{port + 1000};
+    keepalive 32;
+}}
+
+server {{
+    listen {port};
+    server_name {server_name_config};
+    
+    # Security headers
+    add_header X-Frame-Options "SAMEORIGIN" always;
+    add_header X-XSS-Protection "1; mode=block" always;
+    add_header X-Content-Type-Options "nosniff" always;
+    add_header Referrer-Policy "no-referrer-when-downgrade" always;
+    add_header Content-Security-Policy "default-src 'self' http: https: data: blob: 'unsafe-inline'" always;
+    
+    # Gzip compression
+    gzip on;
+    gzip_vary on;
+    gzip_min_length 1024;
+    gzip_proxied expired no-cache no-store private must-revalidate auth;
+    gzip_types text/plain text/css text/xml text/javascript application/x-javascript application/xml+rss;
+    
+    # Proxy to Node.js
+    location / {{
+        proxy_pass http://{server_name}_backend;
+        proxy_http_version 1.1;
+        proxy_set_header Upgrade $http_upgrade;
+        proxy_set_header Connection 'upgrade';
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
+        proxy_cache_bypass $http_upgrade;
+        
+        # Timeouts
+        proxy_connect_timeout 60s;
+        proxy_send_timeout 60s;
+        proxy_read_timeout 60s;
+    }}
+    
+    # Static files (if any)
+    location ~* \\.(jpg|jpeg|png|gif|ico|css|js|woff|woff2|ttf|svg)$ {{
+        root {site_path};
+        expires 1y;
+        add_header Cache-Control "public, immutable";
+    }}
+    
+    # Logging
+    access_log /var/log/nginx/{server_name}_access.log;
+    error_log /var/log/nginx/{server_name}_error.log;
+}}
+"""
+        return config.strip()
+    
+    def _generate_static_config(self, server_name: str, port: int, site_path: str, domain: str = None) -> str:
+        """Generate nginx config for static files"""
+        server_name_config = domain or f"{server_name}.local"
+        
+        config = f"""
+server {{
+    listen {port};
+    server_name {server_name_config};
+    root {site_path};
+    index index.html index.htm;
+    
+    # Security headers
+    add_header X-Frame-Options "SAMEORIGIN" always;
+    add_header X-XSS-Protection "1; mode=block" always;
+    add_header X-Content-Type-Options "nosniff" always;
+    add_header Referrer-Policy "no-referrer-when-downgrade" always;
+    add_header Content-Security-Policy "default-src 'self' http: https: data: blob: 'unsafe-inline'" always;
+    
+    # Gzip compression
+    gzip on;
+    gzip_vary on;
+    gzip_min_length 1024;
+    gzip_proxied expired no-cache no-store private must-revalidate auth;
+    gzip_types text/plain text/css text/xml text/javascript application/x-javascript application/xml+rss;
+    
+    # Static files
+    location / {{
+        try_files $uri $uri/ =404;
+        
+        # Cache static files
+        location ~* \\.(jpg|jpeg|png|gif|ico|css|js|woff|woff2|ttf|svg)$ {{
+            expires 1y;
+            add_header Cache-Control "public, immutable";
+        }}
+    }}
+    
+    # Deny access to hidden files
+    location ~ /\\. {{
+        deny all;
+    }}
+    
+    # Logging
+    access_log /var/log/nginx/{server_name}_access.log;
+    error_log /var/log/nginx/{server_name}_error.log;
+}}
+"""
+        return config.strip()
+    
+    def save_nginx_config(self, server_name: str, config: str) -> str:
+        """Save nginx configuration to file"""
+        config_path = self.config_dir / f"{server_name}.conf"
+        config_path.write_text(config, encoding='utf-8')
+        return str(config_path)
+    
+    def enable_site(self, server_name: str) -> bool:
+        """Enable nginx site by creating symlink"""
+        try:
+            source = self.config_dir / f"{server_name}.conf"
+            target = Path(self.sites_enabled) / f"{server_name}.conf"
+            
+            if target.exists():
+                target.unlink()
+            
+            target.symlink_to(source)
+            return True
+        except Exception as e:
+            app_logger.error(f"Failed to enable nginx site {server_name}: {e}")
+            return False
+    
+    def disable_site(self, server_name: str) -> bool:
+        """Disable nginx site by removing symlink"""
+        try:
+            target = Path(self.sites_enabled) / f"{server_name}.conf"
+            if target.exists():
+                target.unlink()
+            return True
+        except Exception as e:
+            app_logger.error(f"Failed to disable nginx site {server_name}: {e}")
+            return False
+    
+    def reload_nginx(self) -> bool:
+        """Reload nginx configuration"""
+        try:
+            result = subprocess.run(['sudo', 'nginx', '-t'], capture_output=True, text=True)
+            if result.returncode != 0:
+                app_logger.error(f"Nginx config test failed: {result.stderr}")
+                return False
+            
+            result = subprocess.run(['sudo', 'systemctl', 'reload', 'nginx'], 
+                                  capture_output=True, text=True)
+            if result.returncode != 0:
+                # Fallback to nginx -s reload
+                result = subprocess.run(['sudo', 'nginx', '-s', 'reload'], 
+                                      capture_output=True, text=True)
+                if result.returncode != 0:
+                    app_logger.error(f"Failed to reload nginx: {result.stderr}")
+                    return False
+            
+            return True
+        except Exception as e:
+            app_logger.error(f"Failed to reload nginx: {e}")
+            return False
+
+# =========================
+# PHP-FPM Configuration Management
+# =========================
+class PHPFPMManager:
+    def __init__(self):
+        self.pool_dir = Config.PHP_FPM_POOL_DIR
+        self.socket_dir = Config.PHP_FPM_SOCKET_DIR
+    
+    def generate_pool_config(self, server_name: str, site_path: str) -> str:
+        """Generate PHP-FPM pool configuration"""
+        config = f"""[{server_name}]
+user = www-data
+group = www-data
+listen = /var/run/php/php8.1-fpm-{server_name}.sock
+listen.owner = www-data
+listen.group = www-data
+listen.mode = 0660
+
+pm = dynamic
+pm.max_children = 20
+pm.start_servers = 2
+pm.min_spare_servers = 1
+pm.max_spare_servers = 3
+pm.max_requests = 1000
+
+chdir = {site_path}
+
+php_admin_value[error_log] = /var/log/php8.1-fpm-{server_name}.log
+php_admin_flag[log_errors] = on
+php_value[session.save_handler] = files
+php_value[session.save_path] = /var/lib/php/sessions
+php_value[soap.wsdl_cache_dir] = /var/lib/php/wsdlcache
+"""
+        return config.strip()
+    
+    def save_pool_config(self, server_name: str, config: str) -> str:
+        """Save PHP-FPM pool configuration"""
+        config_path = Config.PHP_FPM_CONFIG_DIR / f"{server_name}.conf"
+        config_path.write_text(config, encoding='utf-8')
+        return str(config_path)
+    
+    def enable_pool(self, server_name: str) -> bool:
+        """Enable PHP-FPM pool by creating symlink"""
+        try:
+            source = Config.PHP_FPM_CONFIG_DIR / f"{server_name}.conf"
+            target = Path(self.pool_dir) / f"{server_name}.conf"
+            
+            if target.exists():
+                target.unlink()
+            
+            target.symlink_to(source)
+            return True
+        except Exception as e:
+            app_logger.error(f"Failed to enable PHP-FPM pool {server_name}: {e}")
+            return False
+    
+    def disable_pool(self, server_name: str) -> bool:
+        """Disable PHP-FPM pool by removing symlink"""
+        try:
+            target = Path(self.pool_dir) / f"{server_name}.conf"
+            if target.exists():
+                target.unlink()
+            return True
+        except Exception as e:
+            app_logger.error(f"Failed to disable PHP-FPM pool {server_name}: {e}")
+            return False
+    
+    def reload_php_fpm(self) -> bool:
+        """Reload PHP-FPM configuration"""
+        try:
+            result = subprocess.run(['sudo', 'systemctl', 'reload', 'php8.1-fpm'], 
+                                  capture_output=True, text=True)
+            if result.returncode != 0:
+                app_logger.error(f"Failed to reload PHP-FPM: {result.stderr}")
+                return False
+            return True
+        except Exception as e:
+            app_logger.error(f"Failed to reload PHP-FPM: {e}")
+            return False
+
+# =========================
+# Server Management
+# =========================
+class ServerManager:
+    def __init__(self):
+        self.nginx = NginxManager()
+        self.php_fpm = PHPFPMManager()
+    
+    def create_server(self, name: str, server_type: str, port: int, site_path: str, domain: str = None) -> Tuple[bool, str]:
+        """Create a new server with nginx configuration"""
+        try:
+            # Generate nginx configuration
+            nginx_config = self.nginx.generate_nginx_config(name, server_type, port, site_path, domain)
+            nginx_config_path = self.nginx.save_nginx_config(name, nginx_config)
+            
+            # Enable nginx site
+            if not self.nginx.enable_site(name):
+                return False, "Failed to enable nginx site"
+            
+            # For PHP servers, create PHP-FPM pool
+            if server_type == 'PHP':
+                php_fpm_config = self.php_fpm.generate_pool_config(name, site_path)
+                php_fpm_config_path = self.php_fpm.save_pool_config(name, php_fpm_config)
+                
+                if not self.php_fpm.enable_pool(name):
+                    return False, "Failed to enable PHP-FPM pool"
+                
+                if not self.php_fpm.reload_php_fpm():
+                    return False, "Failed to reload PHP-FPM"
+            
+            # Reload nginx
+            if not self.nginx.reload_nginx():
+                return False, "Failed to reload nginx"
+            
+            # Start backend process for Node.js
+            if server_type == 'Node.js':
+                success, msg = self._start_nodejs_backend(name, site_path, port + 1000)
+                if not success:
+                    return False, f"Failed to start Node.js backend: {msg}"
+            
+            return True, "Server created successfully"
+            
+        except Exception as e:
+            app_logger.error(f"Failed to create server {name}: {e}")
+            return False, f"Error creating server: {e}"
+    
+    def _start_nodejs_backend(self, name: str, site_path: str, port: int) -> Tuple[bool, str]:
+        """Start Node.js backend process"""
+        try:
+            # Create Node.js server script
+            script_path = self._create_nodejs_script(name, site_path, port)
+            
+            # Start process
+            cmd = [get_node_path(), script_path]
+            proc = subprocess.Popen(cmd, cwd=site_path, stdout=subprocess.PIPE, 
+                                  stderr=subprocess.PIPE, text=True, bufsize=1)
+            
+            with LOCK:
+                server_processes[name] = proc
+                log_queues.setdefault(name, queue.Queue())
+            
+            # Start monitoring thread
+            threading.Thread(target=self._monitor_process_io, args=(name, proc), daemon=True).start()
+            
+            return True, "Node.js backend started"
+            
+        except Exception as e:
+            app_logger.error(f"Failed to start Node.js backend for {name}: {e}")
+            return False, str(e)
+    
+    def _create_nodejs_script(self, name: str, site_path: str, port: int) -> str:
+        """Create Node.js server script"""
+        script_path = Path(site_path) / f"server_{name}.js"
+        
+        script_content = f"""
+const http = require('http');
+const fs = require('fs');
+const path = require('path');
+const url = require('url');
+
+const siteRoot = '{site_path}';
+const port = {port};
+
+const server = http.createServer((req, res) => {{
+    const parsed = url.parse(req.url);
+    let filePath = parsed.pathname;
+    
+    if (filePath === '/') filePath = '/index.html';
+    
+    const fullPath = path.join(siteRoot, filePath);
+    
+    // Security check
+    if (!fullPath.startsWith(siteRoot)) {{
+        res.writeHead(403, {{'Content-Type': 'text/plain'}});
+        res.end('Forbidden');
+        return;
+    }}
+    
+    fs.readFile(fullPath, (err, data) => {{
+        if (err) {{
+            res.writeHead(404, {{'Content-Type': 'text/html'}});
+            res.end('<h1>404 - File Not Found</h1>');
+            return;
+        }}
+        
+        const ext = path.extname(fullPath);
+        const mimeTypes = {{
+            '.html': 'text/html',
+            '.css': 'text/css',
+            '.js': 'application/javascript',
+            '.json': 'application/json',
+            '.png': 'image/png',
+            '.jpg': 'image/jpeg',
+            '.gif': 'image/gif',
+            '.svg': 'image/svg+xml',
+            '.ico': 'image/x-icon'
+        }};
+        
+        const contentType = mimeTypes[ext] || 'text/plain';
+        res.writeHead(200, {{'Content-Type': contentType}});
+        res.end(data);
+    }});
+}});
+
+server.listen(port, '127.0.0.1', () => {{
+    console.log(`Node.js server for {name} listening on port ${{port}}`);
+}});
+
+// Graceful shutdown
+process.on('SIGTERM', () => {{
+    console.log('Received SIGTERM, shutting down gracefully');
+    server.close(() => {{
+        process.exit(0);
+    }});
+}});
+
+process.on('SIGINT', () => {{
+    console.log('Received SIGINT, shutting down gracefully');
+    server.close(() => {{
+        process.exit(0);
+    }});
+}});
+"""
+        
+        script_path.write_text(script_content, encoding='utf-8')
+        return str(script_path)
+    
+    def _monitor_process_io(self, name: str, proc: subprocess.Popen):
+        """Monitor process output"""
+        logger = logging.getLogger(f"usa_server_{name}")
+        q = log_queues.setdefault(name, queue.Queue())
+        
+        def push(m, t='info'):
+            try:
+                q.put({'timestamp': now_str(), 'message': m, 'type': t})
+            except:
+                pass
+        
+        try:
+            while True:
+                out = proc.stdout.readline() if proc.stdout else ''
+                err = proc.stderr.readline() if proc.stderr else ''
+                
+                if out:
+                    logger.info(out.rstrip('\n'))
+                    push(out.rstrip('\n'), 'info')
+                
+                if err:
+                    logger.error(err.rstrip('\n'))
+                    push(err.rstrip('\n'), 'error')
+                
+                if proc.poll() is not None:
+                    break
+                
+                time.sleep(0.02)
+        except Exception as e:
+            logger.error(f"Monitor error: {e}")
+        finally:
+            with LOCK:
+                if name in servers:
+                    servers[name]['status'] = 'Stopped'
+            logger.info("Process monitor ended")
+    
+    def delete_server(self, name: str) -> Tuple[bool, str]:
+        """Delete a server and clean up configurations"""
+        try:
+            # Stop process if running
+            if name in server_processes:
+                proc = server_processes[name]
+                try:
+                    proc.terminate()
+                    proc.wait(timeout=5)
+                except:
+                    try:
+                        proc.kill()
+                        proc.wait()
+                    except:
+                        pass
+                
+                with LOCK:
+                    server_processes.pop(name, None)
+                    log_queues.pop(name, None)
+            
+            # Disable nginx site
+            self.nginx.disable_site(name)
+            
+            # Disable PHP-FPM pool if exists
+            if name in servers and servers[name].get('type') == 'PHP':
+                self.php_fpm.disable_pool(name)
+                self.php_fpm.reload_php_fpm()
+            
+            # Remove configuration files
+            nginx_config = Config.NGINX_CONFIG_DIR / f"{name}.conf"
+            if nginx_config.exists():
+                nginx_config.unlink()
+            
+            php_fpm_config = Config.PHP_FPM_CONFIG_DIR / f"{name}.conf"
+            if php_fpm_config.exists():
+                php_fpm_config.unlink()
+            
+            # Reload nginx
+            self.nginx.reload_nginx()
+            
+            return True, "Server deleted successfully"
+            
+        except Exception as e:
+            app_logger.error(f"Failed to delete server {name}: {e}")
+            return False, f"Error deleting server: {e}"
+
+# =========================
+# Persistence
+# =========================
 def save_servers_to_disk():
     try:
         with LOCK:
             minimal = {}
             for name, meta in servers.items():
-                # only persist serializable fields
                 minimal[name] = {
                     'name': meta.get('name'),
                     'folder': meta.get('folder'),
                     'port': meta.get('port'),
                     'type': meta.get('type'),
                     'status': meta.get('status'),
-                    'start_time': meta.get('start_time')
+                    'start_time': meta.get('start_time'),
+                    'domain': meta.get('domain')
                 }
-        with open(PERSIST_FILE, 'w', encoding='utf-8') as f:
+        
+        with open(Config.PERSIST_FILE, 'w', encoding='utf-8') as f:
             json.dump(minimal, f, indent=2)
-    except Exception:
+    except Exception as e:
         app_logger.exception("save_servers_to_disk failed")
 
 def load_servers_from_disk():
-    if not PERSIST_FILE.exists():
+    if not Config.PERSIST_FILE.exists():
         return
+    
     try:
-        with open(PERSIST_FILE, 'r', encoding='utf-8') as f:
+        with open(Config.PERSIST_FILE, 'r', encoding='utf-8') as f:
             data = json.load(f)
+        
         with LOCK:
             for name, meta in data.items():
-                # only register if folder still exists
                 folder = meta.get('folder')
                 if folder and Path(folder).exists():
                     servers[name] = {
                         'name': meta.get('name', name),
                         'folder': folder,
                         'port': meta.get('port'),
-                        'type': meta.get('type', 'HTTP'),
-                        'status': meta.get('status','Stopped'),
-                        'start_time': meta.get('start_time')
+                        'type': meta.get('type', 'Static'),
+                        'status': meta.get('status', 'Stopped'),
+                        'start_time': meta.get('start_time'),
+                        'domain': meta.get('domain')
                     }
-    except Exception:
+    except Exception as e:
         app_logger.exception("load_servers_from_disk failed")
 
-# -------------------------
-# Folder import / upload
-# -------------------------
-def import_folder_recursive(src: str, dest_basename: str = None) -> Tuple[bool,str]:
-    srcp = Path(src)
-    if not srcp.exists() or not srcp.is_dir():
-        return False, "Source not a directory"
-    if not dest_basename:
-        dest_basename = f"{srcp.name}-{int(time.time())}"
-    dest = SITES_FOLDER / secure_name(dest_basename)
-    try:
-        shutil.copytree(srcp, dest, dirs_exist_ok=True)
-        app_logger.info("Imported folder %s -> %s", src, dest)
-        return True, str(dest)
-    except Exception:
-        app_logger.exception("import_folder failed")
-        return False, "Import failed"
-
-def extract_zip_to_sites(zip_path: str, dest_basename: str = None) -> Tuple[bool,str]:
-    tmpdir = tempfile.mkdtemp(prefix='msa_extract_')
-    try:
-        with zipfile.ZipFile(zip_path, 'r') as z:
-            z.extractall(tmpdir)
-        entries = [p for p in Path(tmpdir).iterdir() if not p.name.startswith('__MACOSX')]
-        if len(entries) == 1 and entries[0].is_dir():
-            target_name = dest_basename or entries[0].name
-            dest = SITES_FOLDER / secure_name(target_name)
-            shutil.move(str(entries[0]), str(dest))
-        else:
-            target_name = dest_basename or f"site-{int(time.time())}"
-            dest = SITES_FOLDER / secure_name(target_name)
-            os.makedirs(dest, exist_ok=True)
-            for item in Path(tmpdir).iterdir():
-                shutil.move(str(item), str(dest / item.name))
-        app_logger.info("Extracted zip %s -> %s", zip_path, dest)
-        return True, str(dest)
-    except Exception:
-        app_logger.exception("extract_zip failed")
-        return False, "Extraction failed"
-    finally:
-        try:
-            shutil.rmtree(tmpdir)
-        except:
-            pass
-
-# -------------------------
-# Process management
-# -------------------------
-def _create_node_script(folder_abs: str, port:int, bind_all:bool=True) -> str:
-    bind = '0.0.0.0' if bind_all else '127.0.0.1'
-    script = Path(folder_abs) / f"msa_temp_server_{port}.js"
-    content = f"""
-const http = require('http');
-const fs = require('fs');
-const path = require('path');
-const url = require('url');
-const siteRoot = {json.dumps(folder_abs)};
-const server = http.createServer((req,res) => {{
-  const parsed = url.parse(req.url);
-  let p = parsed.pathname;
-  if (p === '/') p = '/index.html';
-  const full = path.join(siteRoot, p);
-  fs.readFile(full, (err, data) => {{
-    if (err) {{
-      res.writeHead(404, {{'Content-Type':'text/html'}});
-      res.end('<h1>404</h1>');
-      return;
-    }}
-    const ext = path.extname(full);
-    const mimes = {{'.html':'text/html','.css':'text/css','.js':'application/javascript','.json':'application/json','.png':'image/png','.jpg':'image/jpeg','.gif':'image/gif','.svg':'image/svg+xml'}};
-    res.writeHead(200, {{'Content-Type': mimes[ext] || 'text/plain'}});
-    res.end(data);
-  }});
-}});
-server.listen({port}, '{bind}', () => console.log('Node listening http://{bind}:{port}'));
-"""
-    script.write_text(content, encoding='utf-8')
-    return str(script)
-
-def start_server_process(name: str, folder: str, port: int, server_type: str, bind_all: bool = True, health_wait: float = 0.2) -> Tuple[bool,str]:
-    logger = get_server_logger(name)
-    try:
-        folder_abs = str(Path(folder).absolute())
-        if server_type == 'PHP':
-            php = get_php_path()
-            if not check_php_available():
-                return False, "PHP not available"
-            cmd = [php, '-S', f"{'0.0.0.0' if bind_all else '127.0.0.1'}:{int(port)}", '-t', folder_abs, '-d', 'display_errors=1', '-d', 'log_errors=1']
-        elif server_type == 'HTTP':
-            cmd = [sys.executable, '-m', 'http.server', str(int(port)), '--bind', ('0.0.0.0' if bind_all else '127.0.0.1')]
-        elif server_type == 'Node.js':
-            nodeexec = shutil.which('node')
-            if not nodeexec:
-                return False, "Node.js not available"
-            script = _create_node_script(folder_abs, int(port), bind_all)
-            cmd = [nodeexec, script]
-        else:
-            return False, "Unsupported server type"
-
-        proc = subprocess.Popen(cmd, cwd=folder_abs, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True, bufsize=1)
-        with LOCK:
-            server_processes[name] = proc
-            log_queues.setdefault(name, queue.Queue())
-        threading.Thread(target=_monitor_process_io, args=(name, proc), daemon=True).start()
-
-        time.sleep(health_wait)
-        if proc.poll() is not None:
-            try:
-                err = proc.stderr.read() if proc.stderr else ''
-                out = proc.stdout.read() if proc.stdout else ''
-            except:
-                err = out = ''
-            with LOCK:
-                server_processes.pop(name, None)
-                log_queues.pop(name, None)
-            msg = (err.strip() or out.strip() or "Exiting immediately")
-            logger.error("Health-check failed: %s", msg)
-            return False, msg
-
-        logger.info("Started process successfully")
-        return True, "Started"
-    except Exception:
-        app_logger.exception("start_server_process failed")
-        return False, "Failed to start process"
-
-def _monitor_process_io(name: str, proc: subprocess.Popen):
-    logger = get_server_logger(name)
-    q = log_queues.setdefault(name, queue.Queue())
-    def push(m,t='info'):
-        try:
-            q.put({'timestamp': now_str(), 'message': m, 'type': t})
-        except:
-            pass
-    try:
-        while True:
-            out = proc.stdout.readline() if proc.stdout else ''
-            err = proc.stderr.readline() if proc.stderr else ''
-            if out:
-                logger.info(out.rstrip('\n'))
-                push(out.rstrip('\n'), 'info')
-            if err:
-                logger.error(err.rstrip('\n'))
-                push(err.rstrip('\n'), 'error')
-            if proc.poll() is not None:
-                try:
-                    rem = proc.stdout.read() if proc.stdout else ''
-                    for l in rem.splitlines(): logger.info(l); push(l,'info')
-                except: pass
-                try:
-                    rem = proc.stderr.read() if proc.stderr else ''
-                    for l in rem.splitlines(): logger.error(l); push(l,'error')
-                except: pass
-                break
-            time.sleep(0.02)
-    except Exception:
-        logger.exception("monitor io error")
-    finally:
-        with LOCK:
-            if name in servers:
-                servers[name]['status'] = 'Stopped'
-        logger.info("Process monitor ended")
-
-# -------------------------
-# Background monitor & persistence
-# -------------------------
+# =========================
+# Background Monitor
+# =========================
 def monitor_loop():
-    app_logger.info("Background monitor started (interval %s)", RUNTIME_OPTIONS['AUTO_REFRESH_INTERVAL'])
-    interval = float(RUNTIME_OPTIONS['AUTO_REFRESH_INTERVAL'])
+    app_logger.info("Background monitor started")
     while True:
         try:
-            # update server entries: mark Running if process exists & listening; else Stopped
             with LOCK:
                 for name, meta in list(servers.items()):
                     proc = server_processes.get(name)
                     port = meta.get('port')
+                    
                     if proc:
                         if proc.poll() is None:
                             meta['status'] = 'Running'
                         else:
                             meta['status'] = 'Stopped'
                     else:
-                        # if no process, but port is in use by some other program, consider it 'Running (external)'
                         if port and is_port_in_use(port, '0.0.0.0'):
                             meta['status'] = 'Running'
                         else:
                             meta['status'] = 'Stopped'
-            # persist every cycle
+            
             save_servers_to_disk()
-        except Exception:
+        except Exception as e:
             app_logger.exception("monitor_loop error")
-        time.sleep(interval)
+        
+        time.sleep(Config.AUTO_REFRESH_INTERVAL)
 
 # Start monitor thread
-t = threading.Thread(target=monitor_loop, daemon=True)
-t.start()
+monitor_thread = threading.Thread(target=monitor_loop, daemon=True)
+monitor_thread.start()
 
-# -------------------------
-# Discover sites on boot & load persisted servers
-# -------------------------
-def discover_and_load():
-    # load persisted
-    load_servers_from_disk()
-    # scan sites folder and register any non-registered folders
-    for entry in SITES_FOLDER.iterdir():
-        if entry.is_dir():
-            name_base = secure_name(entry.name)
-            with LOCK:
-                if name_base not in servers:
-                    # choose unique name
-                    n = name_base
-                    i = 1
-                    while n in servers:
-                        n = f"{name_base}-{i}"; i += 1
-                    servers[n] = {
-                        'name': n,
-                        'folder': str(entry),
-                        'port': None,
-                        'type': 'HTTP',
-                        'status': 'Stopped',
-                        'start_time': None
-                    }
-    # optionally auto-start discovered (disabled by default)
-    if RUNTIME_OPTIONS['AUTO_START_DISCOVERED']:
-        with LOCK:
-            to_start = [ (n,s) for n,s in servers.items() if s.get('status')!='Running' ]
-        for name, meta in to_start:
-            if not meta.get('port'):
-                p = 8000
-                while p < 65535 and (is_port_in_use(p,'0.0.0.0') or any(s.get('port')==p for s in servers.values())):
-                    p += 1
-                meta['port'] = p
-            success,msg = start_server_process(name, meta['folder'], meta['port'], meta['type'], bind_all=RUNTIME_OPTIONS['BIND_ALL_DEFAULT'])
-            with LOCK:
-                if success:
-                    meta['status'] = 'Running'
-                    meta['start_time'] = now_str()
-                else:
-                    meta['status'] = 'Stopped'
-
-# call discovery/load
-discover_and_load()
-
-# If AUTO_RESTORE_RUNNING True, try to start servers that were previously 'Running' on disk
-if RUNTIME_OPTIONS['AUTO_RESTORE_RUNNING']:
-    with LOCK:
-        to_restore = [ (n,m) for n,m in servers.items() if m.get('status')=='Running' ]
-    for name, meta in to_restore:
-        port = meta.get('port') or 8000
-        while is_port_in_use(port,'0.0.0.0') or any(s.get('port')==port for s in servers.values() if s['name']!=name):
-            port += 1
-            if port > 65535: break
-        success, msg = start_server_process(name, meta['folder'], port, meta.get('type','HTTP'), bind_all=RUNTIME_OPTIONS['BIND_ALL_DEFAULT'])
-        with LOCK:
-            if success:
-                meta['status'] = 'Running'
-                meta['port'] = port
-                meta['start_time'] = now_str()
-            else:
-                meta['status'] = 'Stopped'
-
-# -------------------------
-# Routes
-# -------------------------
+# =========================
+# Flask Routes
+# =========================
 @app.route('/')
 def index():
-    # UI unchanged; front-end will call /api/servers automatically
-    return render_template('index.html', servers=servers, current_folder=str(SITES_FOLDER))
+    return render_template('unified_index.html', servers=servers)
 
 @app.route('/api/servers')
 def api_servers():
-    # returns up-to-date server list (monitor loop keeps fresh)
     with LOCK:
         out = {}
         for name, s in servers.items():
@@ -514,285 +879,80 @@ def api_servers():
                 'port': s.get('port'),
                 'type': s.get('type'),
                 'status': s.get('status'),
-                'start_time': s.get('start_time')
+                'start_time': s.get('start_time'),
+                'domain': s.get('domain')
             }
     return jsonify(out)
-
-@app.route('/api/refresh', methods=['POST'])
-def api_refresh():
-    # manual refresh endpoint (UI can call)
-    try:
-        discover_and_load()
-        return jsonify({'success': True, 'message': 'Refreshed'})
-    except Exception as e:
-        app_logger.exception("api_refresh error")
-        return jsonify({'success': False, 'message': str(e)})
 
 @app.route('/api/add_server', methods=['POST'])
 def api_add_server():
     try:
         data = request.get_json() or {}
-        folder = data.get('folder','').strip()
-        port = data.get('port', 8000)
-        server_type = data.get('type','HTTP')
-        auto_import = bool(data.get('auto_import', True))
-        auto_start = bool(data.get('auto_start', True))
-        if not folder:
-            return jsonify({'success': False, 'message': 'Select folder'})
-
-        folder_p = Path(folder)
-        if folder_p.exists() and folder_p.is_dir() and SITES_FOLDER in folder_p.parents:
-            final = str(folder_p)
-        elif folder_p.exists() and folder_p.is_dir() and auto_import:
-            ok,res = import_folder_recursive(str(folder_p), dest_basename=folder_p.name)
-            if not ok: return jsonify({'success': False, 'message': f'Import failed: {res}'})
-            final = res
-        else:
-            dest = SITES_FOLDER / secure_name(folder_p.name or f"site-{int(time.time())}")
-            os.makedirs(dest, exist_ok=True)
-            final = str(dest)
-
-        try:
-            port = int(port)
-        except:
-            return jsonify({'success': False, 'message': 'Invalid port'})
-        if port<1024 or port>65535: return jsonify({'success': False, 'message':'Port out of range'})
-
-        orig = port
-        while is_port_in_use(port,'0.0.0.0') or any(s.get('port')==port for s in servers.values()):
-            port += 1
-            if port>65535: return jsonify({'success': False, 'message':f'No free port from {orig}'})
-
-        base = secure_name(Path(final).name)
-        with LOCK:
-            server_name = base
-            i=1
-            while server_name in servers:
-                server_name = f"{base}-{i}"; i+=1
-            servers[server_name] = {
-                'name': server_name,
-                'folder': final,
-                'port': port,
-                'type': server_type,
-                'status': 'Stopped',
-                'start_time': None
-            }
-        # attempt start
-        success,msg = start_server_process(server_name, final, port, server_type, bind_all=RUNTIME_OPTIONS['BIND_ALL_DEFAULT'])
-        with LOCK:
-            if success:
-                servers[server_name]['status'] = 'Running'
-                servers[server_name]['start_time'] = now_str()
-            else:
-                servers.pop(server_name, None)
-                save_servers_to_disk()
-                return jsonify({'success': False, 'message': f'Start failed: {msg}'})
-        save_servers_to_disk()
-        return jsonify({'success': True, 'message': f'Started {server_name} on port {port}'})
-    except Exception:
-        app_logger.exception("api_add_server error")
-        return jsonify({'success': False, 'message': 'Error'})
-
-@app.route('/api/stop_server', methods=['POST'])
-def api_stop_server():
-    try:
-        data = request.get_json() or {}
-        name = data.get('name')
-        if not name: return jsonify({'success': False, 'message': 'Missing name'})
-        with LOCK:
-            proc = server_processes.get(name)
-        if not proc:
+        name = data.get('name', '').strip()
+        server_type = data.get('type', 'Static')
+        port = int(data.get('port', 8000))
+        site_path = data.get('site_path', '').strip()
+        domain = data.get('domain', '').strip() or None
+        
+        if not name:
+            return jsonify({'success': False, 'message': 'Server name is required'})
+        
+        if not site_path or not Path(site_path).exists():
+            return jsonify({'success': False, 'message': 'Valid site path is required'})
+        
+        if port < 1000 or port > 65535:
+            return jsonify({'success': False, 'message': 'Port must be between 1000 and 65535'})
+        
+        if is_port_in_use(port):
+            return jsonify({'success': False, 'message': f'Port {port} is already in use'})
+        
+        # Create server
+        server_manager = ServerManager()
+        success, message = server_manager.create_server(name, server_type, port, site_path, domain)
+        
+        if success:
             with LOCK:
-                if name in servers: servers[name]['status'] = 'Stopped'
+                servers[name] = {
+                    'name': name,
+                    'folder': site_path,
+                    'port': port,
+                    'type': server_type,
+                    'status': 'Running',
+                    'start_time': now_str(),
+                    'domain': domain
+                }
             save_servers_to_disk()
-            return jsonify({'success': False, 'message': 'Not running'})
-        try:
-            proc.terminate()
-            try: proc.wait(timeout=5)
-            except subprocess.TimeoutExpired:
-                proc.kill(); proc.wait()
-        except: pass
-        with LOCK:
-            server_processes.pop(name, None)
-            log_queues.pop(name, None)
-            if name in servers: servers[name]['status'] = 'Stopped'
-        save_servers_to_disk()
-        return jsonify({'success': True, 'message': 'Stopped'})
-    except Exception:
-        app_logger.exception("api_stop_server error")
-        return jsonify({'success': False, 'message': 'Error'})
-
-@app.route('/api/start_server', methods=['POST'])
-def api_start_server():
-    try:
-        data = request.get_json() or {}
-        name = data.get('name')
-        if not name or name not in servers:
-            return jsonify({'success': False, 'message': 'Server not found'})
-        with LOCK:
-            meta = servers[name]
-        if meta.get('status') == 'Running':
-            return jsonify({'success': False, 'message': 'Already running'})
-        port = meta.get('port') or 8000
-        while is_port_in_use(port,'0.0.0.0') or any(s.get('port')==port for s in servers.values() if s['name']!=name):
-            port += 1
-            if port>65535: return jsonify({'success': False,'message':'No free port'})
-        success,msg = start_server_process(name, meta['folder'], port, meta.get('type','HTTP'), bind_all=RUNTIME_OPTIONS['BIND_ALL_DEFAULT'])
-        with LOCK:
-            if success:
-                meta['port'] = port
-                meta['status'] = 'Running'
-                meta['start_time'] = now_str()
-                save_servers_to_disk()
-                return jsonify({'success': True, 'message':'Started'})
-            else:
-                return jsonify({'success': False, 'message': msg})
-    except Exception:
-        app_logger.exception("api_start_server error")
-        return jsonify({'success': False, 'message': 'Error'})
+            return jsonify({'success': True, 'message': f'Server {name} created successfully'})
+        else:
+            return jsonify({'success': False, 'message': message})
+            
+    except Exception as e:
+        app_logger.exception("api_add_server error")
+        return jsonify({'success': False, 'message': f'Error: {e}'})
 
 @app.route('/api/delete_server', methods=['POST'])
 def api_delete_server():
     try:
         data = request.get_json() or {}
-        name = data.get('name'); delete_files = bool(data.get('delete_files', RUNTIME_OPTIONS['DELETE_SITE_ON_REMOVE']))
+        name = data.get('name')
+        
         if not name or name not in servers:
-            return jsonify({'success': False, 'message': 'Not found'})
-        with LOCK:
-            proc = server_processes.pop(name, None)
-        if proc:
-            try:
-                proc.terminate(); proc.wait(timeout=3)
-            except:
-                try: proc.kill(); proc.wait()
-                except: pass
-        with LOCK:
-            meta = servers.pop(name, None)
-            log_queues.pop(name, None)
-        if delete_files and meta and meta.get('folder'):
-            try:
-                shutil.rmtree(meta['folder'])
-            except Exception:
-                app_logger.exception("delete site files error")
-                return jsonify({'success': False, 'message': 'Server deleted but failed to delete files'})
-        save_servers_to_disk()
-        return jsonify({'success': True, 'message': 'Deleted'})
-    except Exception:
-        app_logger.exception("api_delete_server error")
-        return jsonify({'success': False, 'message': 'Error'})
-
-@app.route('/api/upload_zip', methods=['POST'])
-def api_upload_zip():
-    try:
-        if 'file' not in request.files:
-            return jsonify({'success': False, 'message': 'No file'})
-        f = request.files['file']
-        if f.filename == '':
-            return jsonify({'success': False, 'message': 'Empty filename'})
-        if not (f.filename.lower().endswith('.zip') or allowed_file(f.filename)):
-            return jsonify({'success': False, 'message': 'Not allowed type'})
-        name = request.form.get('name')
-        auto_register = request.form.get('auto_register','true').lower()!='false'
-        auto_start = request.form.get('auto_start','false').lower()=='true'
-        tmp_fd, tmp_path = tempfile.mkstemp(suffix='.zip')
-        os.close(tmp_fd)
-        f.save(tmp_path)
-        ok,res = extract_zip_to_sites(tmp_path, dest_basename=name)
-        os.remove(tmp_path)
-        if not ok:
-            return jsonify({'success': False, 'message': res})
-        site_path = res
-        registered = None
-        if auto_register:
-            base = secure_name(Path(site_path).name)
-            with LOCK:
-                n = base; i=1
-                while n in servers: n = f"{base}-{i}"; i+=1
-                servers[n] = {'name': n, 'folder': site_path, 'port': None, 'type':'HTTP', 'status':'Stopped','start_time':None}
-                registered = n
-            if auto_start:
-                p = 8000
-                while is_port_in_use(p,'0.0.0.0') or any(s.get('port')==p for s in servers.values()):
-                    p+=1
-                success,msg = start_server_process(registered, site_path, p, 'HTTP', bind_all=RUNTIME_OPTIONS['BIND_ALL_DEFAULT'])
-                if success:
-                    with LOCK:
-                        servers[registered]['port'] = p; servers[registered]['status']='Running'; servers[registered]['start_time']=now_str()
-                else:
-                    return jsonify({'success': False, 'message': f'Imported but failed to auto-start: {msg}', 'registered': registered})
-        save_servers_to_disk()
-        return jsonify({'success': True, 'site_path': site_path, 'registered': registered})
-    except Exception:
-        app_logger.exception("api_upload_zip error")
-        return jsonify({'success': False, 'message': 'Error'})
-
-@app.route('/api/upload_folder', methods=['POST'])
-def api_upload_folder():
-    try:
-        # clients should prefer zip. But support multiple files with folder_name
-        files = request.files.getlist('files[]')
-        folder_name = request.form.get('folder_name') or f"site-{int(time.time())}"
-        if not files:
-            return jsonify({'success': False, 'message': 'No files'})
-        dest = SITES_FOLDER / secure_name(folder_name)
-        os.makedirs(dest, exist_ok=True)
-        for f in files:
-            filename = f.filename
-            if '/' in filename or '\\' in filename:
-                parts = Path(filename).parts
-                subdir = dest.joinpath(*parts[:-1])
-                subdir.mkdir(parents=True, exist_ok=True)
-                f.save(str(subdir / secure_name(parts[-1])))
-            else:
-                f.save(str(dest / secure_name(filename)))
-        # register
-        base = secure_name(dest.name)
-        with LOCK:
-            n = base; i=1
-            while n in servers: n = f"{base}-{i}"; i+=1
-            servers[n] = {'name': n, 'folder': str(dest), 'port': None, 'type':'HTTP', 'status':'Stopped','start_time':None}
-        save_servers_to_disk()
-        return jsonify({'success': True, 'registered': n, 'site_path': str(dest)})
-    except Exception:
-        app_logger.exception("api_upload_folder error")
-        return jsonify({'success': False, 'message': 'Error'})
-
-@app.route('/api/open_browser/<server_name>')
-def api_open_browser(server_name):
-    try:
-        if server_name not in servers:
             return jsonify({'success': False, 'message': 'Server not found'})
-        meta = servers[server_name]
-        port = meta.get('port')
-        if not port:
-            return jsonify({'success': False, 'message': 'Server not started'})
-        # Determine host IP to return (so remote clients can open it)
-        host_ip = request.host.split(':')[0]
-        if host_ip in ('0.0.0.0', ''):
-            # try to find machine IP
-            try:
-                s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-                s.connect(("8.8.8.8", 80))
-                host_ip = s.getsockname()[0]
-                s.close()
-            except Exception:
-                host_ip = request.remote_addr or '127.0.0.1'
-        # if remote open not allowed, return localhost/127.0.0.1
-        if not RUNTIME_OPTIONS['ALLOW_REMOTE_OPEN_BROWSER']:
-            host_ip = '127.0.0.1'
-        url = f"http://{host_ip}:{port}"
-        # if client wants local host to open
-        if request.args.get('open_local','false').lower() == 'true':
-            try:
-                import webbrowser
-                webbrowser.open(url)
-                return jsonify({'success': True, 'url': url, 'message': 'Opening locally'})
-            except Exception as e:
-                return jsonify({'success': False, 'url': url, 'message': f'Failed to open local: {e}'})
-        return jsonify({'success': True, 'url': url})
-    except Exception:
-        app_logger.exception("api_open_browser error")
-        return jsonify({'success': False, 'message': 'Error'})
+        
+        server_manager = ServerManager()
+        success, message = server_manager.delete_server(name)
+        
+        if success:
+            with LOCK:
+                servers.pop(name, None)
+            save_servers_to_disk()
+            return jsonify({'success': True, 'message': f'Server {name} deleted successfully'})
+        else:
+            return jsonify({'success': False, 'message': message})
+            
+    except Exception as e:
+        app_logger.exception("api_delete_server error")
+        return jsonify({'success': False, 'message': f'Error: {e}'})
 
 @app.route('/api/server_logs/<server_name>')
 def api_server_logs(server_name):
@@ -805,91 +965,51 @@ def api_server_logs(server_name):
                     logs.append(q.get_nowait())
             except:
                 pass
+        
         if not logs:
-            logfile = LOGS_FOLDER / f"{server_name}.log"
+            logfile = Config.LOGS_FOLDER / f"{server_name}.log"
             if logfile.exists():
-                lines = tail_lines(str(logfile), n=RUNTIME_OPTIONS['LOG_TAIL_LINES'])
+                lines = tail_lines(str(logfile), n=Config.LOG_TAIL_LINES)
                 logs = [{'timestamp': None, 'message': l, 'type': 'info'} for l in lines]
+        
         return jsonify({'logs': logs})
-    except Exception:
+    except Exception as e:
         app_logger.exception("api_server_logs error")
-        return jsonify({'logs': [], 'error':'Error'})
-
-@app.route('/api/check_php')
-def api_check_php():
-    try:
-        ok = check_php_available()
-        ver = subprocess.run([get_php_path(),'--version'], capture_output=True, text=True).stdout.splitlines()[0] if ok else None
-        return jsonify({'available': ok, 'version': ver})
-    except Exception:
-        app_logger.exception("api_check_php error")
-        return jsonify({'available': False, 'version': None})
-
-@app.route('/api/check_node')
-def api_check_node():
-    try:
-        ok = check_node_available()
-        ver = subprocess.run([get_node_path(),'--version'], capture_output=True, text=True).stdout.strip() if ok else None
-        return jsonify({'available': ok, 'version': ver})
-    except Exception:
-        app_logger.exception("api_check_node error")
-        return jsonify({'available': False, 'version': None})
+        return jsonify({'logs': [], 'error': 'Error'})
 
 @app.route('/api/system_info')
 def api_system_info():
     try:
-        os_name = platform.system(); os_ver = platform.release()
         info = {
-            'os': f"{os_name} {os_ver}",
-            'cpu_cores': (psutil.cpu_count() if PSUTIL_AVAILABLE else None),
+            'os': f"{platform.system()} {platform.release()}",
+            'cpu_cores': psutil.cpu_count() if PSUTIL_AVAILABLE else None,
             'python_version': platform.python_version(),
-            'active_servers': len([s for s in servers.values() if s.get('status')=='Running']),
-            'sites_folder': str(SITES_FOLDER)
+            'active_servers': len([s for s in servers.values() if s.get('status') == 'Running']),
+            'sites_folder': str(Config.SITES_FOLDER),
+            'php_available': check_php_available(),
+            'node_available': check_node_available(),
+            'nginx_available': check_nginx_available()
         }
         return jsonify(info)
-    except Exception:
+    except Exception as e:
         app_logger.exception("api_system_info error")
         return jsonify({})
 
-# -------------------------
-# Graceful shutdown
-# -------------------------
-def shutdown_all():
-    app_logger.info("Shutting down child servers")
-    with LOCK:
-        items = list(server_processes.items())
-    for name, proc in items:
-        try:
-            proc.terminate()
-            proc.wait(timeout=3)
-        except:
-            try:
-                proc.kill()
-            except:
-                pass
-        with LOCK:
-            server_processes.pop(name, None)
-            log_queues.pop(name, None)
-            if name in servers:
-                servers[name]['status'] = 'Stopped'
-    save_servers_to_disk()
-
-# -------------------------
-# Run app
-# -------------------------
+# =========================
+# Main Application
+# =========================
 if __name__ == '__main__':
-    app_logger.info("Modern Server Administrator - production-ready (final)")
-    app_logger.info("Sites: %s, Logs: %s", SITES_FOLDER, LOGS_FOLDER)
-    app_logger.info("PHP available: %s, Node available: %s", check_php_available(), check_node_available())
-    # persist and ensure discovered sites are registered
-    save_servers_to_disk()
-    # Start Flask with binding so UI is reachable from devices
+    app_logger.info("Unified Server Administrator - Production Ready")
+    app_logger.info("Nginx: %s, PHP: %s, Node.js: %s", 
+                   check_nginx_available(), check_php_available(), check_node_available())
+    
+    # Load persisted servers
+    load_servers_from_disk()
+    
     try:
-        app.run(host='0.0.0.0', port=5000, debug=False, threaded=True)
+        app.run(host=Config.HOST, port=Config.PORT, debug=Config.DEBUG, threaded=True)
     except KeyboardInterrupt:
-        shutdown_all()
-        app_logger.info("Goodbye")
-    except Exception:
+        app_logger.info("Shutting down...")
+    except Exception as e:
         app_logger.exception("Fatal app error")
-        shutdown_all()
         sys.exit(1)
